video_filters renamed to filters
